Updates for renames and structure change
