+++ Doxyfile
@@ -950,13 +950,13 @@ WARN_LOGFILE           =
 # Note: If this tag is empty the current directory is searched.
 
 INPUT                  =  docs/moq-api-main.md \
-                          include/moqt/core.h \
-                          include/moqt/api/config.h \
-                          include/moqt/api/client.h \
-                          include/moqt/api/server.h \
-                          include/moqt/api/subscribe_track_handler.h \
-                          include/moqt/api/publish_track_handler.h \
-                          include/moqt/base_track_handler.h
+                          include/moqt/core/transport.h \
+                          include/moqt/config.h \
+                          include/moqt/client.h \
+                          include/moqt/server.h \
+                          include/moqt/subscribe_track_handler.h \
+                          include/moqt/publish_track_handler.h \
+                          include/moqt/core/base_track_handler.h
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
